SourceCode/requirements.txt
langchain==0.3.22
langgraph==0.3.34

langchain-anthropic==0.3.10
langchain-groq==0.2.4
langchain-openai==0.3.9
langchain-deepseek==0.1.3
langchain_ollama
langchain-google-genai
google-ai-generativelanguage

pandas==2.2.3
numpy==2.2.1
requests==2.32.3

python-dotenv==1.1.0

matplotlib==3.10.0
matplotlib-inline==0.1.7
tabulate==0.9.0
colorama==0.4.6

questionary==2.1.0
rich==13.9.4
streamlit==1.41.1
plotly==6.1.2

pytest==8.3.4
black==25.1.0
isort==6.0.1
flake8==7.1.2

# FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-multipart==0.0.9
httpx==0.27.0